Fix the dps chart redesign
