+++ components/unitStats/dps/dpsChart.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useRef, useState } from "react";
 import { useMediaQuery } from "@mantine/hooks";
 
-//import { LevelContext } from './LevelContext.js';
 import {
   Chart as ChartJS,
   CategoryScale,
@@ -37,10 +36,10 @@ import {
 } from "@mantine/core";
 import { UnitSearch } from "./unitSearch";
 import { DpsUnitCustomizing } from "./dpsUnitCustomizing";
-import { EbpsType, getEbpsStats } from "../../src/unitStats/mappingEbps";
+import { EbpsType, getEbpsStats } from "../../../src/unitStats/mappingEbps";
 // import slash from "slash";
-import { getWeaponStats, WeaponType } from "../../src/unitStats/mappingWeapon";
-import { getSbpsStats, SbpsType } from "../../src/unitStats/mappingSbps";
+import { getWeaponStats, WeaponType } from "../../../src/unitStats/mappingWeapon";
+import { getSbpsStats, SbpsType } from "../../../src/unitStats/mappingSbps";
 import { IconAdjustments } from "@tabler/icons-react";
 import {
   CustomizableUnit,
@@ -48,17 +47,19 @@ import {
   getWeaponDPSData,
   mapCustomizableUnit,
   updateHealth,
-} from "../../src/unitStats/dpsCommon";
-import { getFactionIcon } from "../../src/unitStats";
-import config from "../../config";
+} from "../../../src/unitStats/dpsCommon";
+import { getFactionIcon } from "../../../src/unitStats";
+import config from "../../../config";
 import {
   AnalyticsDPSExplorerPatchSelection,
   AnalyticsDPSExplorerSquadSelection,
-} from "../../src/firebase/analytics";
+} from "../../../src/firebase/analytics";
 
 // let unitSelectionList :  CustomizableUnit[] = [];
 let unitSelectionList1: CustomizableUnit[] = [];
 let unitSelectionList2: CustomizableUnit[] = [];
+
+import classes from "../DPSChart.module.css";
 
 // function hexToRgbA(hex: string, opacity: string) {
 //   let c: any;
@@ -204,7 +205,7 @@ const generateFilterButtons = (
         <ActionIcon
           key={faction + index}
           size="sm"
-          variant={unitFilter.includes(faction) ? "gradient" : "trannsparent"}
+          variant={unitFilter.includes(faction) ? "gradient" : "transparent"}
           onClick={() => callback(faction, index, unitFilter, unitSelectionList)}
         >
           <Image src={source} alt={"Filter"}></Image>
@@ -494,10 +495,10 @@ export const DpsChart = (props: IDPSProps) => {
         <Space h="xl" />
         <>
           <Grid>
-            <Grid.Col span={{md: 6, lg: 6}}>
+            <Grid.Col span={{ md: 6, lg: 6 }}>
               <Grid>
                 <Grid.Col span={6}>
-                   <Group wrap="nowrap">
+                  <Group wrap="nowrap">
                     {generateFilterButtons(unitFilter1, toggleFilter, 1, unitSelectionList1)}
                   </Group>
                 </Grid.Col>
@@ -519,6 +520,8 @@ export const DpsChart = (props: IDPSProps) => {
                         onChange={(value) => onPatchUnitChange(value as string, 1)}
                         data={patchList}
                         defaultValue={config.latestPatch}
+                        withCheckIcon={false}
+                        allowDeselect={false}
                       />
                     </Tooltip>
                   </Flex>
@@ -543,16 +546,7 @@ export const DpsChart = (props: IDPSProps) => {
                   </Center>
                 )}
                 {activeData[0] && (
-                  <Box
-                    sx={(theme) => ({
-                      backgroundColor:
-                        theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.white,
-                      border: "solid 2px " + theme.colors.blue[4],
-                      textAlign: "left",
-                      padding: theme.spacing.xs,
-                      borderRadius: theme.radius.md,
-                    })}
-                  >
+                  <Box className={classes.unitBoxLeft}>
                     <DpsUnitCustomizing
                       key={activeData[0].id + "0." + patchUnit1}
                       unit={activeData[0]}
@@ -566,15 +560,14 @@ export const DpsChart = (props: IDPSProps) => {
               </div>
             </Grid.Col>
 
-            <Grid.Col span={{md: 6, lg: 6}}>
-              {/* <SimpleGrid cols={2}> */}
+            <Grid.Col span={{ md: 6, lg: 6 }}>
               <Grid>
-                <Grid.Col span={{md: 6, lg: 6}}>
-                   <Group wrap="nowrap">
+                <Grid.Col span={{ md: 6, lg: 6 }}>
+                  <Group wrap="nowrap">
                     {generateFilterButtons(unitFilter2, toggleFilter, 2, unitSelectionList2)}
                   </Group>
                 </Grid.Col>
-                <Grid.Col span={{md: 6, lg: 6}}>
+                <Grid.Col span={{ md: 6, lg: 6 }}>
                   <Flex
                     // mih={50}
                     // gap="xs"
@@ -594,6 +587,8 @@ export const DpsChart = (props: IDPSProps) => {
                           onChange={(value) => onPatchUnitChange(value as string, 2)}
                           data={patchList}
                           defaultValue={config.latestPatch}
+                          withCheckIcon={false}
+                          allowDeselect={false}
                         />
                       </Tooltip>
                     </Group>
@@ -620,16 +615,7 @@ export const DpsChart = (props: IDPSProps) => {
                   </Center>
                 )}
                 {activeData[1] && (
-                  <Box
-                    sx={(theme) => ({
-                      backgroundColor:
-                        theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.white,
-                      border: "solid 2px " + theme.colors.red[6],
-                      textAlign: "left",
-                      padding: theme.spacing.xs,
-                      borderRadius: theme.radius.md,
-                    })}
-                  >
+                  <Box className={classes.unitBoxRight}>
                     <DpsUnitCustomizing
                       key={activeData[1].id + "1." + patchUnit2}
                       unit={activeData[1]}
@@ -648,20 +634,8 @@ export const DpsChart = (props: IDPSProps) => {
 
       <Space h="sm" />
       <Container size="md">
-        <LoadingOverlay visible={patchChangeIndex > 0} overlayBlur={1} />
-        <Box
-          sx={(theme) => ({
-            backgroundColor:
-              theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.white,
-            border:
-              theme.colorScheme === "dark"
-                ? "solid 1px " + theme.colors.dark[4]
-                : "solid 1px " + theme.colors.gray[4],
-            textAlign: "left",
-            // padding: theme.spacing.xs,
-            borderRadius: theme.radius.md,
-          })}
-        >
+        <LoadingOverlay visible={patchChangeIndex > 0} />
+        <Box className={classes.chartBox}>
           {/* { patchChangeIndex > 0 &&
             <Container size={'md'}>
                 <Loader />
@@ -677,8 +651,8 @@ export const DpsChart = (props: IDPSProps) => {
           infantry. Values do not necessarily reflect the average time to kill (ttk) in game.
         </Text>
         <Text c={"dimmed"} pl={5} fs="italic">
-          ** Area and balistic DPS (Eg. by Mortar, Tank Guns..) are experimental approximations
-          respecting scatter, penetration and target size. Some informations like box sizes of
+          ** Area and ballistic DPS (Eg. by Mortar, Tank Guns..) are experimental approximations
+          respecting scatter, penetration and target size. Some information like box sizes of
           units are not accessible. Also, squads formations and densities are unknown. The
           calculation assumes every model within the model cap limit to be hit which allows a
           rough performance comparison but do not necessarily reflect the average time to kill in
